+++ lfc/core/src/main/java/org/lflang/generator/LFGenerator.java
@@ -2,20 +2,17 @@ package org.lflang.generator;
 
 import com.google.inject.Inject;
 import com.google.inject.Injector;
-import java.io.IOException;
+
 import java.nio.file.Path;
-import java.util.Arrays;
+
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.xtext.generator.AbstractGenerator;
 import org.eclipse.xtext.generator.IFileSystemAccess2;
 import org.eclipse.xtext.generator.IGeneratorContext;
-import org.eclipse.xtext.util.RuntimeIOException;
 import org.lflang.FileConfig;
 import org.lflang.MessageReporter;
 import org.lflang.ast.ASTUtils;
 import org.lflang.generator.uc.UcFileConfig;
-import org.lflang.generator.uc.UcGenerator;
-import org.lflang.generator.uc.UcNonFederatedGenerator;
 import org.lflang.scoping.LFGlobalScopeProvider;
 import org.lflang.target.Target;
