+++ lfc/core/src/main/kotlin/org/lflang/generator/uc/UcCmakeGenerator.kt
@@ -18,61 +18,71 @@ import java.time.LocalDateTime
 import kotlin.io.path.name
 import kotlin.math.max
 
-class UcCmakeGenerator(private val mainDef: Instantiation, private val targetConfig: TargetConfig, private val fileConfig: FileConfig, private val numEvents: Int, private val numReactions: Int) {
-    private val S = '$' // a little trick to escape the dollar sign with $S
-    private val platform = targetConfig.get(PlatformProperty.INSTANCE).platform
-    private val includeFiles = targetConfig.get(CmakeIncludeProperty.INSTANCE)?.map { fileConfig.srcPath.resolve(it).toUnixString() }
-    private val mainTarget = fileConfig.name
+abstract class UcCmakeGenerator(private val targetConfig: TargetConfig, private val fileConfig: UcFileConfig, private val numEvents: Int, private val numReactions: Int) {
+    protected val S = '$' // a little trick to escape the dollar sign with $S
+    private val minCmakeVersion = "3.10"
+    protected val includeFiles = targetConfig.get(CmakeIncludeProperty.INSTANCE)?.map { fileConfig.srcPath.resolve(it).toUnixString() }
+    protected val platform = targetConfig.get(PlatformProperty.INSTANCE).platform
+    abstract val mainTarget: String
+    abstract fun generateCmake(sources: List<Path>): String
 
-
-    fun generateCmake(sources: List<Path>) =
-        if (platform == PlatformType.Platform.NATIVE) {
-            generateCmakePosix(sources)
-        } else {
-            generateCmakeEmbedded(sources)
-        }
-
-    fun generateCmakeEmbedded(sources: List<Path>) = with(PrependOperator) {
-        """
-            |# This file is generated by LFC. It is meant to be included in
-            |# an existing CMake project. 
-            |
+    protected fun generateCmakeCommon(sources: List<Path>, compileDefs: List<String>) = with(PrependOperator) {
+        """ |
             |set(LFC_GEN_SOURCES
         ${" |    "..sources.filterNot{it.name == "lf_main.c"}.joinWithLn { "$S{CMAKE_CURRENT_LIST_DIR}/${it.toUnixString()}"}}
+            |)
+            |set(LFC_GEN_COMPILE_DEFS
+        ${" |    "..compileDefs.joinWithLn { it }}
             |)
             |set(LFC_GEN_MAIN "$S{CMAKE_CURRENT_LIST_DIR}/lf_main.c")
             |set(REACTOR_UC_PATH $S{CMAKE_CURRENT_LIST_DIR}/reactor-uc)
             |set(LFC_GEN_INCLUDE_DIRS $S{CMAKE_CURRENT_LIST_DIR})
             |set(REACTION_QUEUE_SIZE ${max(numReactions, 1)} CACHE STRING "Size of the reaction queue")
             |set(EVENT_QUEUE_SIZE ${max(numEvents, 1)} CACHE STRING "Size of the event queue")
-            |
         """.trimMargin()
     }
 
-    fun generateCmakePosix(sources: List<Path>) = with(PrependOperator) {
+    protected fun generateCmakePosix(sources: List<Path>, compileDefs: List<String>) = with(PrependOperator) {
         """
-            |cmake_minimum_required(VERSION 3.10)
+            |cmake_minimum_required(VERSION $minCmakeVersion)
             |project(${mainTarget} LANGUAGES C)
-            |set(PLATFORM POSIX CACHE STRING "Target platform")
-            |set(REACTION_QUEUE_SIZE ${max(numReactions, 1)} CACHE STRING "Size of the reaction queue")
-            |set(EVENT_QUEUE_SIZE ${max(numEvents, 1)} CACHE STRING "Size of the event queue")
-            |set(CMAKE_BUILD_TYPE ${targetConfig.getOrDefault(BuildTypeProperty.INSTANCE)})
-            |
             |set(LF_MAIN_TARGET ${mainTarget})
-            |set(SOURCES
-        ${" |    "..sources.joinWithLn { it.toUnixString() }}
-            |)
-            |add_executable($S{LF_MAIN_TARGET} $S{SOURCES})
+            |set(CMAKE_BUILD_TYPE ${targetConfig.getOrDefault(BuildTypeProperty.INSTANCE)})
+            |set(PLATFORM POSIX CACHE STRING "Target platform")
+        ${" |"..generateCmakeCommon(sources, compileDefs)}
+            |add_executable($S{LF_MAIN_TARGET} $S{LFC_GEN_SOURCES} $S{LFC_GEN_MAIN})
             |install(TARGETS $S{LF_MAIN_TARGET}
             |        RUNTIME DESTINATION $S{CMAKE_INSTALL_BINDIR}
             |        OPTIONAL
             |)
-            |add_compile_definitions(LF_LOG_LEVEL_ALL=${targetConfig.getOrDefault(LoggingProperty.INSTANCE).ordinal})
-            |
-            |add_subdirectory(reactor-uc $S{CMAKE_CURRENT_LIST_DIR})
+            |add_compile_definitions("LF_LOG_LEVEL_ALL=LF_LOG_LEVEL_${targetConfig.getOrDefault(LoggingProperty.INSTANCE).name.uppercase()}")
+            |add_compile_definitions($S{LFC_GEN_COMPILE_DEFS})
+            |add_subdirectory($S{REACTOR_UC_PATH})
             |target_link_libraries($S{LF_MAIN_TARGET} PRIVATE reactor-uc)
-            |target_include_directories($S{LF_MAIN_TARGET} PRIVATE $S{CMAKE_CURRENT_LIST_DIR})
-            ${" |"..(includeFiles?.joinWithLn { "include(\"$it\")" } ?: "")}
+            |target_include_directories($S{LF_MAIN_TARGET} PRIVATE $S{LFC_GEN_INCLUDE_DIRS})
+        ${" |"..(includeFiles?.joinWithLn { "include(\"$it\")" } ?: "")}
         """.trimMargin()
     }
+}
+
+class UcCmakeGeneratorNonFederated(private val mainDef: Instantiation, targetConfig: TargetConfig, fileConfig: UcFileConfig, numEvents: Int, numReactions: Int): UcCmakeGenerator(targetConfig, fileConfig, numEvents, numReactions) {
+    override val mainTarget = fileConfig.name
+
+    override fun generateCmake(sources: List<Path>) =
+        if (platform == PlatformType.Platform.NATIVE) {
+            generateCmakePosix(sources, emptyList())
+        } else {
+            generateCmakeCommon(sources, emptyList())
+        }
+}
+
+class UcCmakeGeneratorFederated(private val federate: UcFederate, targetConfig: TargetConfig, fileConfig: UcFileConfig, numEvents: Int, numReactions: Int): UcCmakeGenerator(targetConfig, fileConfig, numEvents, numReactions) {
+    override val mainTarget = federate.codeType
+
+    override fun generateCmake(sources: List<Path>) =
+        if (platform == PlatformType.Platform.NATIVE) {
+            generateCmakePosix(sources, federate.getCompileDefs())
+        } else {
+            generateCmakeCommon(sources, federate.getCompileDefs())
+        }
 }
