+++ lfc/core/src/main/kotlin/org/lflang/generator/uc/UcConnectionGenerator.kt
@@ -7,16 +7,79 @@ import org.lflang.generator.uc.UcInstanceGenerator.Companion.isAFederate
 import org.lflang.generator.uc.UcReactorGenerator.Companion.codeType
 import org.lflang.lf.*
 
+/**
+ * This generator creates code for configuring the connections between reactors. This is perhaps the most complicated
+ * part of the code-generator. This generator handles both federated and non-federated programs
+ */
 class UcConnectionGenerator(
-    private val reactor: Reactor,
-    private val currentFederate: UcFederate?,
-    private val allFederates: List<UcFederate>
+    private val reactor: Reactor, // The reactor to generator connections for
+    private val currentFederate: UcFederate?, // The federate to generate connections for. If set then `reactor` should be the top-level reactor.
+    private val allFederates: List<UcFederate> // A list of all the federates in the program. Only used for federated code-gen.
 ) {
 
+    /** A list containing all non-federated gruoped connections within this reactor. */
     private val nonFederatedConnections: List<UcGroupedConnection>
+
+    /** A list containing all federated connection bundles (each containing grouped connections) within this reactor, that
+     * has the current federate as a src or dest.
+     */
     private val federatedConnectionBundles: List<UcFederatedConnectionBundle>
+
     private val isFederated = currentFederate != null
 
+    /**
+     * Given a LF connection and possibly the list of federates of the program. Create all the ConnectionChannels
+     * found within the LF Connection. This must handle multiports, banks, iterated connections and federated connections.
+     */
+    private fun parseConnectionChannels(conn: Connection, federates: List<UcFederate>): List<UcConnectionChannel> {
+        val res = mutableListOf<UcConnectionChannel>()
+        val rhsPorts = conn.rightPorts.map {  getChannelQueue(it, federates) }
+        var rhsPortIndex = 0
+
+        var lhsPorts = conn.leftPorts.map {  getChannelQueue(it, federates) }
+        var lhsPortIndex = 0
+
+        // Keep parsing out connections until we are out of right-hand-side (rhs) ports
+        while (rhsPortIndex < rhsPorts.size) {
+            // First get the current lhs and rhs port and UcGroupedConnection that we are working with
+            val lhsPort = lhsPorts[lhsPortIndex]
+            val rhsPort = rhsPorts[rhsPortIndex]
+            if (rhsPort.channelsLeft() > lhsPort.channelsLeft()) {
+                val rhsChannelsToAdd = rhsPort.takeChannels(lhsPort.channelsLeft())
+                val lhsChannelsToAdd = lhsPort.takeRemainingChannels()
+                lhsChannelsToAdd.zip(rhsChannelsToAdd)
+                    .forEach { res.add(UcConnectionChannel(it.first, it.second, conn)) }
+                lhsPortIndex += 1
+            } else if (rhsPort.channelsLeft() < lhsPort.channelsLeft()) {
+                val numRhsChannelsToAdd = rhsPort.channelsLeft()
+                val rhsChannelsToAdd = rhsPort.takeRemainingChannels()
+                val lhsChannelsToAdd = lhsPort.takeChannels(numRhsChannelsToAdd)
+                lhsChannelsToAdd.zip(rhsChannelsToAdd)
+                    .forEach { res.add(UcConnectionChannel(it.first, it.second, conn)) }
+                rhsPortIndex += 1
+            } else {
+                lhsPort.takeRemainingChannels().zip(rhsPort.takeRemainingChannels())
+                    .forEach { res.add(UcConnectionChannel(it.first, it.second, conn)) }
+                rhsPortIndex += 1
+                lhsPortIndex += 1
+            }
+
+            // If we are out of lhs variables, but not rhs, then there should be an iterated connection.
+            // We handle it by resetting the lhsChannels variable and index and continuing until
+            // we have been through all rhs channels.
+            if (lhsPortIndex >= lhsPorts.size && rhsPortIndex < rhsPorts.size) {
+                assert(conn.isIterated)
+                lhsPorts = conn.leftPorts.map {getChannelQueue(it, federates)}
+                lhsPortIndex = 0
+            }
+        }
+        return res
+    }
+
+    /**
+     * Given a list of ConnectionChannels, group them together. How they are grouepd depends on
+     * whether we are dealing with federated or non-federated reactors.
+     */
     private fun groupConnections(channels: List<UcConnectionChannel>): List<UcGroupedConnection> {
         val res = mutableListOf<UcGroupedConnection>()
         val channels = HashSet(channels)
@@ -65,68 +128,27 @@ class UcConnectionGenerator(
         return res
     }
 
-    private fun getUcPorts(portVarRefs: List<VarRef>, federates: List<UcFederate>): List<UcPort> {
-        return portVarRefs.map { c ->
-            if (c.container?.isAFederate ?: false) {
-                val federates = allFederates.filter { it.inst == c.container }
-                UcPort(c, federates)
+    /** Given a port VarRef, and the list of federates. Create a channel queue. I.e. create
+     * all the UcChannels and associate them with the correct federates.
+     */
+    private fun getChannelQueue(portVarRef: VarRef, federates: List<UcFederate>): UcChannelQueue {
+        return if (portVarRef.container?.isAFederate ?: false) {
+                val federates = allFederates.filter { it.inst == portVarRef.container }
+                UcChannelQueue(portVarRef, federates)
             } else {
-                UcPort(c, emptyList())
+                UcChannelQueue(portVarRef, emptyList())
             }
-        }
 
     }
 
-    private fun parseConnectionChannels(conn: Connection, federates: List<UcFederate>): List<UcConnectionChannel> {
-        val res = mutableListOf<UcConnectionChannel>()
-        val rhsPorts = getUcPorts(conn.rightPorts, federates)
-        var rhsPortIndex = 0
-
-        var lhsPorts = getUcPorts(conn.leftPorts, federates)
-        var lhsPortIndex = 0
-
-        // Keep parsing out connections until we are out of right-hand-side (rhs) ports
-        while (rhsPortIndex < rhsPorts.size) {
-            // First get the current lhs and rhs port and UcGroupedConnection that we are working with
-            val lhsPort = lhsPorts[lhsPortIndex]
-            val rhsPort = rhsPorts[rhsPortIndex]
-            if (rhsPort.channelsLeft() > lhsPort.channelsLeft()) {
-                val rhsChannelsToAdd = rhsPort.takeChannels(lhsPort.channelsLeft())
-                val lhsChannelsToAdd = lhsPort.takeRemainingChannels()
-                lhsChannelsToAdd.zip(rhsChannelsToAdd)
-                    .forEach { res.add(UcConnectionChannel(it.first, it.second, conn)) }
-                lhsPortIndex += 1
-            } else if (rhsPort.channelsLeft() < lhsPort.channelsLeft()) {
-                val numRhsChannelsToAdd = rhsPort.channelsLeft()
-                val rhsChannelsToAdd = rhsPort.takeRemainingChannels()
-                val lhsChannelsToAdd = lhsPort.takeChannels(numRhsChannelsToAdd)
-                lhsChannelsToAdd.zip(rhsChannelsToAdd)
-                    .forEach { res.add(UcConnectionChannel(it.first, it.second, conn)) }
-                rhsPortIndex += 1
-            } else {
-                lhsPort.takeRemainingChannels().zip(rhsPort.takeRemainingChannels())
-                    .forEach { res.add(UcConnectionChannel(it.first, it.second, conn)) }
-                rhsPortIndex += 1
-                lhsPortIndex += 1
-            }
-
-            // If we are out of lhs variables, but not rhs, then there should be an iterated connection.
-            // We handle it by resetting the lhsChannels variable and index and continuing until
-            // we have been through all rhs channels.
-            if (lhsPortIndex >= lhsPorts.size && rhsPortIndex < rhsPorts.size) {
-                assert(conn.isIterated)
-                lhsPorts = getUcPorts(conn.leftPorts, federates)
-                lhsPortIndex = 0
-            }
-        }
-        return res
-    }
 
     companion object {
         private val Connection.delayString
             get(): String = this.delay.orNever().toCCode()
 
+        /** Whether we have initialized the UcFederates with NetworkInterfaces. This is only done once. */
         private var federateInterfacesInitialized = false
+        /** A global list of FederatedConnectionBundles. It is computed once and reused when code-generating  */
         private var allFederatedConnectionBundles: List<UcFederatedConnectionBundle> = emptyList()
 
         private fun createFederatedConnectionBundles(groupedConnections: List<UcGroupedConnection>) {
