+++ lfc/core/src/main/kotlin/org/lflang/generator/uc/UcGenerator.kt
@@ -7,29 +7,25 @@ import org.eclipse.xtext.xbase.lib.IteratorExtensions
 import org.lflang.allInstantiations
 import org.lflang.allReactions
 import org.lflang.generator.*
-import org.lflang.generator.GeneratorUtils.canGenerate
 import org.lflang.generator.uc.UcInstanceGenerator.Companion.width
 import org.lflang.generator.uc.UcReactorGenerator.Companion.hasStartup
-import org.lflang.isGeneric
 import org.lflang.lf.Instantiation
-import org.lflang.lf.LfFactory
 import org.lflang.lf.Reactor
 import org.lflang.reactor
 import org.lflang.scoping.LFGlobalScopeProvider
 import org.lflang.target.Target
 import org.lflang.target.property.*
-import org.lflang.target.property.type.PlatformType
-import org.lflang.util.FileUtil
 import java.nio.file.Path
 
+/** Creates either a Federated or NonFederated generator depending on the type of LF program */
 fun createUcGenerator(context: LFGeneratorContext, scopeProvider: LFGlobalScopeProvider): UcGenerator {
     val nodes: Iterable<EObject> = IteratorExtensions.toIterable(context.getFileConfig().resource.getAllContents());
     for (reactor in nodes.filterIsInstance<Reactor>()) {
         if (reactor.isFederated) {
-            return UcFederatedGenerator(context, scopeProvider)
+            return UcGeneratorFederated(context, scopeProvider)
         }
     }
-    return UcNonFederatedGenerator(context, scopeProvider)
+    return UcGeneratorNonFederated(context, scopeProvider)
 }
 
 @Suppress("unused")
@@ -44,10 +40,13 @@ abstract class UcGenerator(
     val fileConfig: UcFileConfig = context.fileConfig as UcFileConfig
     val platform = targetConfig.get(PlatformProperty.INSTANCE)
 
+    // Contains the maximum number of pending events required by each reactor.
+    // Is updated as reactors are analyzed and code-generated.
     val maxNumPendingEvents = mutableMapOf<Reactor, Int>()
 
-
-    private fun _totalNumEventsAndReactions(inst: Instantiation): Triple<Int, Int, Boolean> {
+    // Compute the total number of events and reactions within an instance (and its children)
+    // Also returns whether there is any startup event within the instance.
+    private fun totalNumEventsAndReactions(inst: Instantiation): Triple<Int, Int, Boolean> {
         var numEvents = 0
         var numReactions = 0
         var hasStartup = false
@@ -55,7 +54,7 @@ abstract class UcGenerator(
         remaining.addAll(inst.reactor.allInstantiations)
         while (remaining.isNotEmpty()) {
             val child = remaining.removeFirst()
-            val childRes = _totalNumEventsAndReactions(child)
+            val childRes = totalNumEventsAndReactions(child)
 
             numEvents += childRes.first * child.width
             numReactions += childRes.second * child.width
@@ -67,11 +66,12 @@ abstract class UcGenerator(
         return Triple(numEvents, numReactions, hasStartup)
     }
 
+    // Compute the total number of events and reactions for a top-level reactor.
     fun totalNumEventsAndReactions(main: Reactor): Pair<Int, Int> {
         val res = MutablePair(maxNumPendingEvents[main]!!, main.allReactions.size)
         var hasStartup = main.hasStartup
         for (inst in main.allInstantiations) {
-            val childRes = _totalNumEventsAndReactions(inst)
+            val childRes = totalNumEventsAndReactions(inst)
             res.left += childRes.first * inst.width
             res.right += childRes.second * inst.width
             hasStartup = hasStartup or childRes.third
@@ -86,9 +86,8 @@ abstract class UcGenerator(
         const val MINIMUM_CMAKE_VERSION = "3.5"
     }
 
-    // Returns a possibly empty list of the federates in the current program
-
-    fun getAllFederates(): List<Instantiation> {
+    // Returns a possibly empty list of the federates in the current program.
+    protected fun getAllFederates(): List<Instantiation> {
         val res = mutableListOf<Instantiation>()
         for (reactor in reactors) {
             if (reactor.isFederated) {
@@ -98,8 +97,8 @@ abstract class UcGenerator(
         return res
     }
 
-
-    fun getAllInstantiatedReactors(top: Reactor): List<Reactor> {
+    // Returns a list of all instantiated reactors within a top-level reactor.
+    protected fun getAllInstantiatedReactors(top: Reactor): List<Reactor> {
         val res = mutableListOf<Reactor>()
         for (inst in top.allInstantiations) {
             res.add(inst.reactor)
@@ -107,69 +106,8 @@ abstract class UcGenerator(
         }
         return res.distinct()
     }
-}
 
-class UcNonFederatedGenerator(
-    context: LFGeneratorContext, scopeProvider: LFGlobalScopeProvider
-) : UcGenerator(context, scopeProvider) {
-
-
-    fun doGenerateReactor(
-        resource: Resource,
-        context: LFGeneratorContext,
-        srcGenPath: Path,
-    ): GeneratorResult.Status {
-        if (!canGenerate(errorsOccurred(), mainDef, messageReporter, context)) return GeneratorResult.Status.FAILED
-
-        // generate header and source files for all reactors
-        getAllInstantiatedReactors(mainDef.reactor).map { generateReactorFiles(it, srcGenPath) }
-
-        generateReactorFiles(mainDef.reactor, srcGenPath)
-
-        for (r in getAllImportedResources(resource)) {
-            val generator = UcPreambleGenerator(r, fileConfig, scopeProvider)
-            val headerFile = fileConfig.getPreambleHeaderPath(r);
-            val preambleCodeMap = CodeMap.fromGeneratedCode(generator.generateHeader())
-            codeMaps[srcGenPath.resolve(headerFile)] = preambleCodeMap
-            FileUtil.writeToFile(preambleCodeMap.generatedCode, srcGenPath.resolve(headerFile), true)
-        }
-        return GeneratorResult.Status.GENERATED
-    }
-
-
-    override fun doGenerate(resource: Resource, context: LFGeneratorContext) {
-        super.doGenerate(resource, context)
-
-        if (getAllFederates().isNotEmpty()) {
-            context.errorReporter.nowhere().error("Federated program detected in non-federated generator")
-            return
-        }
-
-        val res = doGenerateReactor(
-            resource,
-            context,
-            fileConfig.srcGenPath,
-        )
-        if (res == GeneratorResult.Status.GENERATED) {
-            // generate platform specific files
-            val platformGenerator = UcStandalonePlatformGenerator(this, fileConfig.srcGenPath)
-            platformGenerator.generatePlatformFiles()
-
-            if (platform.platform == PlatformType.Platform.NATIVE) {
-                if (platformGenerator.doCompile(context)) {
-                    context.finish(GeneratorResult.Status.COMPILED, codeMaps)
-                } else {
-                    context.finish(GeneratorResult.Status.FAILED, codeMaps)
-                }
-            } else {
-                context.finish(GeneratorResult.GENERATED_NO_EXECUTABLE.apply(context, codeMaps))
-            }
-        } else {
-            context.finish(GeneratorResult.Status.FAILED, codeMaps)
-        }
-    }
-
-    private fun getAllImportedResources(resource: Resource): Set<Resource> {
+    protected fun getAllImportedResources(resource: Resource): Set<Resource> {
         val resources: MutableSet<Resource> = scopeProvider.getImportedResources(resource)
         val importedRresources = resources.subtract(setOf(resource))
         resources.addAll(importedRresources.map { getAllImportedResources(it) }.flatten())
@@ -177,184 +115,6 @@ class UcNonFederatedGenerator(
         return resources
     }
 
-    fun generateReactorFiles(reactor: Reactor, srcGenPath: Path) {
-        val generator = UcReactorGenerator(reactor, fileConfig, messageReporter)
-        maxNumPendingEvents.set(reactor, generator.getMaxNumPendingEvents())
-
-        val headerFile = fileConfig.getReactorHeaderPath(reactor)
-        val sourceFile = fileConfig.getReactorSourcePath(reactor)
-        val reactorCodeMap = CodeMap.fromGeneratedCode(generator.generateSource())
-        if (!reactor.isGeneric) ucSources.add(sourceFile)
-        codeMaps[srcGenPath.resolve(sourceFile)] = reactorCodeMap
-        val headerCodeMap = CodeMap.fromGeneratedCode(generator.generateHeader())
-        codeMaps[srcGenPath.resolve(headerFile)] = headerCodeMap
-
-        FileUtil.writeToFile(headerCodeMap.generatedCode, srcGenPath.resolve(headerFile), true)
-        FileUtil.writeToFile(reactorCodeMap.generatedCode, srcGenPath.resolve(sourceFile), true)
-    }
-
     override fun getTarget() = Target.UC
-    override fun getTargetTypes(): TargetTypes = UcTypes
-}
-
-class UcFederatedGenerator(
-    context: LFGeneratorContext, scopeProvider: LFGlobalScopeProvider
-) : UcGenerator(context, scopeProvider) {
-
-    private val nonFederatedGenerator = UcNonFederatedGenerator(context, scopeProvider)
-    val federates = mutableListOf<UcFederate>()
-
-    fun totalNumEventsAndReactionsFederated(federate: UcFederate): Pair<Int, Int> {
-        val eventsFromFederatedConncetions = maxNumPendingEvents[mainDef.reactor]!!
-        val eventsAndReactionsInFederate = nonFederatedGenerator.totalNumEventsAndReactions(federate.inst.reactor)
-        return Pair(
-            eventsFromFederatedConncetions + eventsAndReactionsInFederate.first,
-            eventsAndReactionsInFederate.second
-        )
-    }
-
-    fun doGenerateFederate(
-        resource: Resource,
-        context: LFGeneratorContext,
-        srcGenPath: Path,
-        federate: UcFederate
-    ): GeneratorResult.Status {
-        if (!canGenerate(
-                errorsOccurred(),
-                federate.inst,
-                messageReporter,
-                context
-            )
-        ) return GeneratorResult.Status.FAILED
-
-        // generate all core files
-        generateFiles(federate, srcGenPath, getAllImportedResources(resource))
-        return GeneratorResult.Status.GENERATED
-    }
-
-    private fun clearStateFromPreviousFederate() {
-        codeMaps.clear()
-        ucSources.clear()
-        maxNumPendingEvents.clear()
-        nonFederatedGenerator.maxNumPendingEvents.clear()
-        nonFederatedGenerator.codeMaps.clear()
-        nonFederatedGenerator.ucSources.clear()
-    }
-
-    private fun createMainDef() {
-        for (reactor in reactors) {
-            if (reactor.isFederated) {
-                this.mainDef = LfFactory.eINSTANCE.createInstantiation()
-                this.mainDef.name = reactor.name
-                this.mainDef.reactorClass = reactor
-            }
-        }
-    }
-
-    override fun doGenerate(resource: Resource, context: LFGeneratorContext) {
-        super.doGenerate(resource, context)
-        createMainDef()
-        for (inst in getAllFederates()) {
-            for (bankIdx in 0..<inst.width) {
-                federates.add(UcFederate(inst, bankIdx))
-            }
-        }
-        for (ucFederate in federates) {
-            clearStateFromPreviousFederate()
-
-            val srcGenPath = if (ucFederate.isBank) {
-                fileConfig.srcGenPath.resolve("${ucFederate.inst.name}_${ucFederate.bankIdx}")
-            } else {
-                fileConfig.srcGenPath.resolve(ucFederate.inst.name)
-            }
-
-            val platformGenerator = UcFederatedPlatformGenerator(this, srcGenPath, ucFederate)
-            val res = doGenerateFederate(
-                ucFederate.inst.eResource()!!,
-                context,
-                srcGenPath,
-                ucFederate
-            )
-
-            if (res == GeneratorResult.Status.FAILED) {
-                context.unsuccessfulFinish()
-                return
-            } else {
-                // generate platform specific files
-                platformGenerator.generatePlatformFiles()
-
-                if (platform.platform == PlatformType.Platform.NATIVE) {
-                    if (!platformGenerator.doCompile(context)) {
-                        context.finish(GeneratorResult.Status.FAILED, codeMaps)
-                        return
-                    }
-                }
-            }
-        }
-        if (platform.platform == PlatformType.Platform.NATIVE) {
-            context.finish(GeneratorResult.Status.COMPILED, codeMaps)
-        } else {
-            context.finish(GeneratorResult.Status.GENERATED, codeMaps)
-        }
-        return
-    }
-
-    private fun getAllImportedResources(resource: Resource): Set<Resource> {
-        val resources: MutableSet<Resource> = scopeProvider.getImportedResources(resource)
-        val importedRresources = resources.subtract(setOf(resource))
-        resources.addAll(importedRresources.map { getAllImportedResources(it) }.flatten())
-        resources.add(resource)
-        return resources
-    }
-
-    private fun generateFederateFiles(federate: UcFederate, srcGenPath: Path) {
-        // First thing is that we need to also generate the top-level federate reactor files
-        nonFederatedGenerator.generateReactorFiles(federate.inst.reactor, srcGenPath)
-
-        // Then we generate a reactor which wraps around the top-level reactor in the federate.
-        val generator = UcFederateGenerator(federate, federates, fileConfig, messageReporter)
-        val top = federate.inst.eContainer() as Reactor
-
-        // Record the number of events and reactions in this reactor
-        maxNumPendingEvents.set(top, generator.getMaxNumPendingEvents())
-
-        val headerFile = srcGenPath.resolve("lf_federate.h")
-        val sourceFile = srcGenPath.resolve("lf_federate.c")
-        val federateCodeMap = CodeMap.fromGeneratedCode(generator.generateSource())
-        ucSources.add(sourceFile)
-        codeMaps[srcGenPath.resolve(sourceFile)] = federateCodeMap
-        val headerCodeMap = CodeMap.fromGeneratedCode(generator.generateHeader())
-        codeMaps[srcGenPath.resolve(headerFile)] = headerCodeMap
-
-        FileUtil.writeToFile(headerCodeMap.generatedCode, srcGenPath.resolve(headerFile), true)
-        FileUtil.writeToFile(federateCodeMap.generatedCode, srcGenPath.resolve(sourceFile), true)
-    }
-
-    private fun generateFiles(federate: UcFederate, srcGenPath: Path, resources: Set<Resource>) {
-        // generate header and source files for all reactors
-        getAllInstantiatedReactors(federate.inst.reactor).map {
-            nonFederatedGenerator.generateReactorFiles(
-                it,
-                srcGenPath
-            )
-        }
-
-        generateFederateFiles(federate, srcGenPath)
-
-        // Collect the info on the generated sources
-        ucSources.addAll(nonFederatedGenerator.ucSources)
-        codeMaps.putAll(nonFederatedGenerator.codeMaps)
-
-        for (r in resources) {
-            val generator = UcPreambleGenerator(r, fileConfig, scopeProvider)
-            val headerFile = fileConfig.getPreambleHeaderPath(r);
-            val preambleCodeMap = CodeMap.fromGeneratedCode(generator.generateHeader())
-            codeMaps[srcGenPath.resolve(headerFile)] = preambleCodeMap
-            FileUtil.writeToFile(preambleCodeMap.generatedCode, srcGenPath.resolve(headerFile), true)
-        }
-    }
-
-    override fun getTarget() = Target.UC
-
     override fun getTargetTypes(): TargetTypes = UcTypes
 }
