+++ lfc/core/src/main/kotlin/org/lflang/generator/uc/UcMakeGenerator.kt
@@ -5,16 +5,18 @@ import org.lflang.target.TargetConfig
 import org.lflang.generator.PrependOperator
 import org.lflang.joinWithLn
 import org.lflang.lf.Reactor
-import org.lflang.target.property.BuildTypeProperty
 import org.lflang.toUnixString
 import java.nio.file.Path
-import java.time.LocalDateTime
 import kotlin.io.path.name
 import kotlin.math.max
 
-class UcMakeGenerator(private val main: Reactor, private val targetConfig: TargetConfig, private val fileConfig: FileConfig, private val numEvents: Int, private val numReactions: Int) {
+abstract class UcMakeGenerator() {
+    abstract fun generateMake(sources: List<Path>): String
+}
+
+class UcMakeGeneratorNonFederated(private val main: Reactor, private val targetConfig: TargetConfig, private val fileConfig: FileConfig, private val numEvents: Int, private val numReactions: Int): UcMakeGenerator() {
     private val S = '$' // a little trick to escape the dollar sign with $S
-    fun generateMake(sources: List<Path>) = with(PrependOperator) {
+    override fun generateMake(sources: List<Path>) = with(PrependOperator) {
         val sources = sources.filterNot { it.name=="lf_main.c" }
         """
             | # Makefile generated for ${fileConfig.name}
