+++ test/lf/src/FederatedAttr.lf
@@ -1,6 +1,6 @@
 target uC {
   platform: Native,
-  timeout: 1 msec
+  timeout: 1 sec
 }
 
 reactor Src(id: int = 0) {
